--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged
@@ -1,198 +1,0 @@
-AA	Codon1	Codon2	Paired_total	Total_codon1	Total_codon2
-Ala	GCA	GCA	15509	49169	49169
-Ala	GCA	GCC	9744	49169	36509
-Ala	GCA	GCG	5723	49169	18758
-Ala	GCA	GCT	16225	49169	60888
-Ala	GCC	GCA	9705	36509	49169
-Ala	GCC	GCC	8333	36509	36509
-Ala	GCC	GCG	3868	36509	18758
-Ala	GCC	GCT	13109	36509	60888
-Ala	GCG	GCA	5705	18758	49169
-Ala	GCG	GCC	3908	18758	36509
-Ala	GCG	GCG	2300	18758	18758
-Ala	GCG	GCT	6081	18758	60888
-Ala	GCT	GCA	16259	60888	49169
-Ala	GCT	GCC	13044	60888	36509
-Ala	GCT	GCG	6002	60888	18758
-Ala	GCT	GCT	23172	60888	60888
-Arg	AGA	AGA	30245	63169	63169
-Arg	AGA	AGG	12611	63169	28460
-Arg	AGA	CGA	3789	63169	9715
-Arg	AGA	CGC	3118	63169	8258
-Arg	AGA	CGG	2140	63169	5792
-Arg	AGA	CGT	8181	63169	19143
-Arg	AGG	AGA	12433	28460	63169
-Arg	AGG	AGG	6167	28460	28460
-Arg	AGG	CGA	2092	28460	9715
-Arg	AGG	CGC	1695	28460	8258
-Arg	AGG	CGG	1222	28460	5792
-Arg	AGG	CGT	3490	28460	19143
-Arg	CGA	AGA	3913	9715	63169
-Arg	CGA	AGG	2054	9715	28460
-Arg	CGA	CGA	907	9715	9715
-Arg	CGA	CGC	667	9715	8258
-Arg	CGA	CGG	544	9715	5792
-Arg	CGA	CGT	1191	9715	19143
-Arg	CGC	AGA	3094	8258	63169
-Arg	CGC	AGG	1631	8258	28460
-Arg	CGC	CGA	599	8258	9715
-Arg	CGC	CGC	664	8258	8258
-Arg	CGC	CGG	524	8258	5792
-Arg	CGC	CGT	1264	8258	19143
-Arg	CGG	AGA	2179	5792	63169
-Arg	CGG	AGG	1225	5792	28460
-Arg	CGG	CGA	474	5792	9715
-Arg	CGG	CGC	425	5792	8258
-Arg	CGG	CGG	366	5792	5792
-Arg	CGG	CGT	773	5792	19143
-Arg	CGT	AGA	8385	19143	63169
-Arg	CGT	AGG	3406	19143	28460
-Arg	CGT	CGA	1260	19143	9715
-Arg	CGT	CGC	1202	19143	8258
-Arg	CGT	CGG	672	19143	5792
-Arg	CGT	CGT	3297	19143	19143
-Gly	GGA	GGA	8457	33972	33972
-Gly	GGA	GGC	6884	33972	29617
-Gly	GGA	GGG	4279	33972	18369
-Gly	GGA	GGT	12662	33972	67526
-Gly	GGC	GGA	6776	29617	33972
-Gly	GGC	GGC	6180	29617	29617
-Gly	GGC	GGG	3681	29617	18369
-Gly	GGC	GGT	11597	29617	67526
-Gly	GGG	GGA	4442	18369	33972
-Gly	GGG	GGC	3684	18369	29617
-Gly	GGG	GGG	2412	18369	18369
-Gly	GGG	GGT	6993	18369	67526
-Gly	GGT	GGA	12611	67526	33972
-Gly	GGT	GGC	11446	67526	29617
-Gly	GGT	GGG	7074	67526	18369
-Gly	GGT	GGT	33684	67526	67526
-Ile	ATA	ATA	17537	56141	56141
-Ile	ATA	ATC	12815	56141	51253
-Ile	ATA	ATT	23895	56141	91241
-Ile	ATC	ATA	12861	51253	56141
-Ile	ATC	ATC	14044	51253	51253
-Ile	ATC	ATT	22614	51253	91241
-Ile	ATT	ATA	23544	91241	56141
-Ile	ATT	ATC	22697	91241	51253
-Ile	ATT	ATT	41955	91241	91241
-Leu	CTA	CTA	5965	40807	40807
-Leu	CTA	CTC	2347	40807	17519
-Leu	CTA	CTG	4517	40807	32446
-Leu	CTA	CTT	5289	40807	38489
-Leu	CTA	TTA	11155	40807	80235
-Leu	CTA	TTG	10611	40807	79741
-Leu	CTC	CTA	2315	17519	40807
-Leu	CTC	CTC	1451	17519	17519
-Leu	CTC	CTG	2103	17519	32446
-Leu	CTC	CTT	2488	17519	38489
-Leu	CTC	TTA	4293	17519	80235
-Leu	CTC	TTG	4369	17519	79741
-Leu	CTG	CTA	4565	32446	40807
-Leu	CTG	CTC	2081	32446	17519
-Leu	CTG	CTG	4185	32446	32446
-Leu	CTG	CTT	4266	32446	38489
-Leu	CTG	TTA	8204	32446	80235
-Leu	CTG	TTG	8451	32446	79741
-Leu	CTT	CTA	5163	38489	40807
-Leu	CTT	CTC	2517	38489	17519
-Leu	CTT	CTG	3976	38489	32446
-Leu	CTT	CTT	5708	38489	38489
-Leu	CTT	TTA	10331	38489	80235
-Leu	CTT	TTG	9849	38489	79741
-Leu	TTA	CTA	11266	80235	40807
-Leu	TTA	CTC	4445	80235	17519
-Leu	TTA	CTG	8393	80235	32446
-Leu	TTA	CTT	10237	80235	38489
-Leu	TTA	TTA	23244	80235	80235
-Leu	TTA	TTG	20816	80235	79741
-Leu	TTG	CTA	10567	79741	40807
-Leu	TTG	CTC	4181	79741	17519
-Leu	TTG	CTG	8498	79741	32446
-Leu	TTG	CTT	9547	79741	38489
-Leu	TTG	TTA	21155	79741	80235
-Leu	TTG	TTG	23978	79741	79741
-Pro	CCA	CCA	22189	53715	53715
-Pro	CCA	CCC	7373	53715	21058
-Pro	CCA	CCG	6096	53715	16552
-Pro	CCA	CCT	15317	53715	41144
-Pro	CCC	CCA	7432	21058	53715
-Pro	CCC	CCC	3581	21058	21058
-Pro	CCC	CCG	2636	21058	16552
-Pro	CCC	CCT	6336	21058	41144
-Pro	CCG	CCA	5929	16552	53715
-Pro	CCG	CCC	2754	16552	21058
-Pro	CCG	CCG	2160	16552	16552
-Pro	CCG	CCT	4909	16552	41144
-Pro	CCT	CCA	15618	41144	53715
-Pro	CCT	CCC	6211	41144	21058
-Pro	CCT	CCG	4740	41144	16552
-Pro	CCT	CCT	12547	41144	41144
-Ser	AGC	AGC	4132	30389	30389
-Ser	AGC	AGT	5480	30389	44212
-Ser	AGC	TCA	5667	30389	58166
-Ser	AGC	TCC	4194	30389	42870
-Ser	AGC	TCG	2899	30389	26681
-Ser	AGC	TCT	7056	30389	71104
-Ser	AGT	AGC	5488	44212	30389
-Ser	AGT	AGT	7912	44212	44212
-Ser	AGT	TCA	8910	44212	58166
-Ser	AGT	TCC	6232	44212	42870
-Ser	AGT	TCG	4001	44212	26681
-Ser	AGT	TCT	10442	44212	71104
-Ser	TCA	AGC	5918	58166	30389
-Ser	TCA	AGT	9079	58166	44212
-Ser	TCA	TCA	13273	58166	58166
-Ser	TCA	TCC	8552	58166	42870
-Ser	TCA	TCG	5712	58166	26681
-Ser	TCA	TCT	14334	58166	71104
-Ser	TCC	AGC	4290	42870	30389
-Ser	TCC	AGT	6442	42870	44212
-Ser	TCC	TCA	8459	42870	58166
-Ser	TCC	TCC	7202	42870	42870
-Ser	TCC	TCG	3962	42870	26681
-Ser	TCC	TCT	11482	42870	71104
-Ser	TCG	AGC	3026	26681	30389
-Ser	TCG	AGT	4005	26681	44212
-Ser	TCG	TCA	5644	26681	58166
-Ser	TCG	TCC	4133	26681	42870
-Ser	TCG	TCG	2905	26681	26681
-Ser	TCG	TCT	6407	26681	71104
-Ser	TCT	AGC	6774	71104	30389
-Ser	TCT	AGT	10200	71104	44212
-Ser	TCT	TCA	14782	71104	58166
-Ser	TCT	TCC	11480	71104	42870
-Ser	TCT	TCG	6546	71104	26681
-Ser	TCT	TCT	19700	71104	71104
-Thr	ACA	ACA	17506	54938	54938
-Thr	ACA	ACC	10284	54938	37813
-Thr	ACA	ACG	7851	54938	24772
-Thr	ACA	ACT	17315	54938	60844
-Thr	ACC	ACA	10436	37813	54938
-Thr	ACC	ACC	8390	37813	37813
-Thr	ACC	ACG	4731	37813	24772
-Thr	ACC	ACT	12750	37813	60844
-Thr	ACG	ACA	7756	24772	54938
-Thr	ACG	ACC	4726	24772	37813
-Thr	ACG	ACG	3674	24772	24772
-Thr	ACG	ACT	7654	24772	60844
-Thr	ACT	ACA	17192	60844	54938
-Thr	ACT	ACC	13016	60844	37813
-Thr	ACT	ACG	7571	60844	24772
-Thr	ACT	ACT	20832	60844	60844
-Val	GTA	GTA	8820	37107	37107
-Val	GTA	GTC	6609	37107	33983
-Val	GTA	GTG	7180	37107	32562
-Val	GTA	GTT	12940	37107	64798
-Val	GTC	GTA	6493	33983	37107
-Val	GTC	GTC	7295	33983	33983
-Val	GTC	GTG	5997	33983	32562
-Val	GTC	GTT	12867	33983	64798
-Val	GTG	GTA	7205	32562	37107
-Val	GTG	GTC	6058	32562	33983
-Val	GTG	GTG	6817	32562	32562
-Val	GTG	GTT	11264	32562	64798
-Val	GTT	GTA	12894	64798	37107
-Val	GTT	GTC	12790	64798	33983
-Val	GTT	GTG	11210	64798	32562
-Val	GTT	GTT	25333	64798	64798
