--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED_effect_matrix.txt
@@ -1,50 +1,0 @@
-            GCA         GCC         GCG         GCT
-GCA  0.10492170 -0.06881776  0.06538898 -0.06459173
-GCC -0.06507576  0.07679236 -0.01712442  0.01067053
-GCG  0.06875042 -0.02718456  0.12586218 -0.09487573
-GCT -0.06654781  0.01570684 -0.08296223  0.07654097
-             AGA         AGG         CGA         CGC         CGG         CGT
-AGA  0.072659577 -0.02129149 -0.09764097 -0.16062166 -0.15716802 -0.01869187
-AGG -0.007279574  0.07750796  0.05133003 -0.01788913  0.05169297 -0.11525915
-CGA -0.126236039  0.07078014  0.35999718  0.05663616  0.19212917 -0.04118665
-CGC -0.154110644  0.02064863  0.17658818  0.37795378  0.25748215  0.07605859
-CGG -0.172253129  0.04911739  0.36818200  0.55040152  0.54397516 -0.14227701
-CGT -0.042566275 -0.09343936 -0.09369310  0.13156245 -0.01336329  0.27325467
-            GGA         GGC         GGG        GGT
-GGA  0.14616408  0.05337719  0.11338309 -0.1401354
-GGC  0.07016656  0.10199352  0.05917016 -0.1048118
-GGG  0.07252729  0.05830764  0.11809568 -0.1079664
-GGT -0.13665801 -0.09300214 -0.11818053  0.1554584
-            ATA          ATC          ATT
-ATA  0.14364516 -0.081304506 -0.055272947
-ATC -0.08459041  0.098875579 -0.002402359
-ATT -0.04118871 -0.006050445  0.035857311
-             CTA         CTC         CTG          CTT         TTA         TTG
-CTA  0.060694673 -0.04113760  0.02092591 -0.026625185  0.01887404 -0.03842174
-CTC -0.027883343  0.39990338  0.08405391  0.105316240 -0.06362918 -0.11378631
-CTG  0.010191092  0.09551435  0.17712453 -0.057247137 -0.04535593 -0.02742487
-CTT -0.002870541  0.09258117  0.01151502  0.140932588 -0.01842951 -0.07891885
-TTA  0.008835429 -0.09564906 -0.06685334 -0.009416356  0.06913471 -0.02092317
-TTG -0.034417814 -0.07393743 -0.03280390 -0.049782317 -0.03661247  0.11660322
-             AGC         AGT         TCA         TCC         TCG          TCT
-AGC  0.254119220  0.14490344 -0.06157260 -0.07700632  0.04607239 -0.121288967
-AGT  0.143234487  0.13453340 -0.01044489 -0.04733896 -0.04836242 -0.090553185
-TCA -0.101374096 -0.02886485  0.09961847 -0.04915950  0.01935870  0.001800617
-TCC -0.097660722 -0.07839435 -0.03870576  0.09839230  0.01279407  0.055615359
-TCG  0.002169156 -0.04931287  0.03164013 -0.02910958  0.14380834 -0.032855884
-TCT -0.084708437 -0.06897611 -0.02856108  0.05579926 -0.05339255  0.092168104
-            CCA         CCC          CCG           CCT
-CCA  0.07250121 -0.08368621 -0.069992149 -0.0144593339
-CCC -0.09096050  0.12621498  0.101913640 -0.0002552812
-CCG -0.04379695  0.05470020  0.099522239 -0.0293278377
-CCT -0.03345330  0.01986516  0.005280516  0.0336609209
-            ACA         ACC         ACG         ACT
-ACA  0.07483160 -0.06906516  0.05609522 -0.04690763
-ACC -0.08262420  0.08737385 -0.06504501  0.04837786
-ACG  0.06903089 -0.06405585  0.10947163 -0.06916306
-ACT -0.04008874  0.02695280 -0.05895840  0.04278468
-            GTA         GTC         GTG         GTT
-GTA  0.12355820 -0.09683621  0.04593753 -0.05939151
-GTC -0.08070083  0.10800199 -0.03972603  0.01879289
-GTG  0.04230832 -0.04939534  0.12774218 -0.06809506
-GTT -0.05603584  0.02492635 -0.06360595  0.05828367
