--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED_effect_vector.txt
@@ -1,198 +1,0 @@
-"Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED"
-0.104921701888251
--0.0688177572888358
-0.0653889808394691
--0.0645917293309439
--0.0650757575499656
-0.0767923606771531
--0.0171244197931674
-0.0106705308585436
-0.0687504184652553
--0.0271845587922137
-0.125862183931156
--0.0948757310444828
--0.0665478078845295
-0.015706837551721
--0.0829622326693603
-0.0765409665237561
-0.0726595774723123
--0.0212914875019273
--0.0976409657738499
--0.160621659067065
--0.157168022049154
--0.0186918734740361
--0.00727957442988851
-0.0775079584868218
-0.0513300263608309
--0.0178891333167876
-0.0516929663137883
--0.115259153460923
--0.12623603867036
-0.0707801436936991
-0.359997176745521
-0.0566361589934048
-0.192129165569057
--0.0411866540335954
--0.15411064414063
-0.0206486321447241
-0.176588177042739
-0.377953784479803
-0.257482147087888
-0.0760585916291342
--0.1722531285843
-0.0491173917024076
-0.368181995927357
-0.550401517821301
-0.543975158782316
--0.14227701141219
--0.0425662751211794
--0.0934393557189139
--0.0936930991698508
-0.131562445773066
--0.0133632884250336
-0.273254670986774
-0.146164083083823
-0.0533771947759265
-0.113383087202063
--0.140135377486621
-0.0701665597457907
-0.101993517159654
-0.0591701609242743
--0.104811803748051
-0.0725272917914517
-0.0583076445065834
-0.118095682354466
--0.107966402046631
--0.13665800886017
--0.0930021394431371
--0.118180527214036
-0.155458375795056
-0.143645161785697
--0.0813045058661671
--0.0552729472251948
--0.0845904084188578
-0.0988755791083539
--0.00240235922844972
--0.0411887136402494
--0.00605044506287885
-0.0358573114227829
-0.060694673300244
--0.0411375962714988
-0.020925910037705
--0.0266251847323358
-0.0188740420590131
--0.0384217362952973
--0.0278833427426382
-0.399903383049219
-0.0840539085406802
-0.105316240422842
--0.0636291794889333
--0.11378631302906
-0.0101910921446471
-0.0955143535132392
-0.17712453133286
--0.057247136672282
--0.0453559321564203
--0.0274248672918068
--0.00287054078042293
-0.0925811705093491
-0.0115150188521239
-0.140932588407342
--0.0184295070167115
--0.0789188529506892
-0.00883542864976827
--0.0956490590654647
--0.0668533381879271
--0.00941635606033492
-0.0691347084782484
--0.0209231747797454
--0.0344178143114791
--0.0739374316249617
--0.0328039013277312
--0.0497823172422057
--0.0366124701590725
-0.116603221244518
-0.254119219501185
-0.144903442330671
--0.0615725958205222
--0.0770063175049681
-0.0460723927678094
--0.121288967419996
-0.143234486875379
-0.134533397499429
--0.0104448924691821
--0.0473389602707081
--0.0483624172623195
--0.0905531848155133
--0.101374096065377
--0.028864852065251
-0.099618468611429
--0.0491595006892021
-0.0193586961057754
-0.00180061738241349
--0.0976607215887731
--0.0783943496440629
--0.0387057630800396
-0.0983922980403901
-0.0127940653810512
-0.0556153591727642
-0.00216915619097141
--0.0493128667831562
-0.031640126179339
--0.0291095845536596
-0.143808342971038
--0.0328558835899142
--0.0847084372771619
--0.068976113317999
--0.0285610844568046
-0.0557992642876025
--0.0533925521174121
-0.0921681043299867
-0.0725012130923806
--0.0836862118444973
--0.0699921493170756
--0.0144593338838623
--0.0909605005287242
-0.126214981791138
-0.101913639928946
--0.000255281179636255
--0.0437969543324158
-0.0547002014715696
-0.0995222391049838
--0.029327837708489
--0.033453298572104
-0.0198651647795564
-0.00528051575717876
-0.0336609208838984
-0.0748316010699172
--0.069065161554672
-0.0560952190679225
--0.0469076310450297
--0.0826241971472064
-0.087373848565119
--0.0650450101158224
-0.0483778583379641
-0.0690308876872434
--0.0640558491024029
-0.109471630883217
--0.0691630601975962
--0.0400887407831951
-0.0269528037652922
--0.0589584021598733
-0.0427846847087729
-0.123558197078949
--0.0968362065338968
-0.0459375260081441
--0.0593915065106629
--0.0807008299680462
-0.108001991863246
--0.0397260283217286
-0.0187928851503963
-0.0423083187700868
--0.0493953436522625
-0.127742184340468
--0.0680950576654202
--0.0560358379283371
-0.0249263528717865
--0.0636059526800439
-0.0582836730532565
