--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED_rscpu_vector.txt
@@ -1,198 +1,0 @@
-"Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED"
-1.10492170188825
-0.931182242711164
-1.06538898083947
-0.935408270669056
-0.934924242450034
-1.07679236067715
-0.982875580206833
-1.01067053085854
-1.06875041846526
-0.972815441207786
-1.12586218393116
-0.905124268955517
-0.93345219211547
-1.01570683755172
-0.91703776733064
-1.07654096652376
-1.07265957747231
-0.978708512498073
-0.90235903422615
-0.839378340932935
-0.842831977950846
-0.981308126525964
-0.992720425570112
-1.07750795848682
-1.05133002636083
-0.982110866683212
-1.05169296631379
-0.884740846539077
-0.87376396132964
-1.0707801436937
-1.35999717674552
-1.0566361589934
-1.19212916556906
-0.958813345966405
-0.84588935585937
-1.02064863214472
-1.17658817704274
-1.3779537844798
-1.25748214708789
-1.07605859162913
-0.8277468714157
-1.04911739170241
-1.36818199592736
-1.5504015178213
-1.54397515878232
-0.85772298858781
-0.957433724878821
-0.906560644281086
-0.906306900830149
-1.13156244577307
-0.986636711574966
-1.27325467098677
-1.14616408308382
-1.05337719477593
-1.11338308720206
-0.859864622513379
-1.07016655974579
-1.10199351715965
-1.05917016092427
-0.895188196251949
-1.07252729179145
-1.05830764450658
-1.11809568235447
-0.892033597953369
-0.86334199113983
-0.906997860556863
-0.881819472785964
-1.15545837579506
-1.1436451617857
-0.918695494133833
-0.944727052774805
-0.915409591581142
-1.09887557910835
-0.99759764077155
-0.958811286359751
-0.993949554937121
-1.03585731142278
-1.06069467330024
-0.958862403728501
-1.02092591003771
-0.973374815267664
-1.01887404205901
-0.961578263704703
-0.972116657257362
-1.39990338304922
-1.08405390854068
-1.10531624042284
-0.936370820511067
-0.88621368697094
-1.01019109214465
-1.09551435351324
-1.17712453133286
-0.942752863327718
-0.95464406784358
-0.972575132708193
-0.997129459219577
-1.09258117050935
-1.01151501885212
-1.14093258840734
-0.981570492983288
-0.921081147049311
-1.00883542864977
-0.904350940934535
-0.933146661812073
-0.990583643939665
-1.06913470847825
-0.979076825220255
-0.965582185688521
-0.926062568375038
-0.967196098672269
-0.950217682757794
-0.963387529840928
-1.11660322124452
-1.25411921950118
-1.14490344233067
-0.938427404179478
-0.922993682495032
-1.04607239276781
-0.878711032580004
-1.14323448687538
-1.13453339749943
-0.989555107530818
-0.952661039729292
-0.951637582737681
-0.909446815184487
-0.898625903934623
-0.971135147934749
-1.09961846861143
-0.950840499310798
-1.01935869610578
-1.00180061738241
-0.902339278411227
-0.921605650355937
-0.96129423691996
-1.09839229804039
-1.01279406538105
-1.05561535917276
-1.00216915619097
-0.950687133216844
-1.03164012617934
-0.97089041544634
-1.14380834297104
-0.967144116410086
-0.915291562722838
-0.931023886682001
-0.971438915543195
-1.0557992642876
-0.946607447882588
-1.09216810432999
-1.07250121309238
-0.916313788155503
-0.930007850682924
-0.985540666116138
-0.909039499471276
-1.12621498179114
-1.10191363992895
-0.999744718820364
-0.956203045667584
-1.05470020147157
-1.09952223910498
-0.970672162291511
-0.966546701427896
-1.01986516477956
-1.00528051575718
-1.0336609208839
-1.07483160106992
-0.930934838445328
-1.05609521906792
-0.95309236895497
-0.917375802852794
-1.08737384856512
-0.934954989884178
-1.04837785833796
-1.06903088768724
-0.935944150897597
-1.10947163088322
-0.930836939802404
-0.959911259216805
-1.02695280376529
-0.941041597840127
-1.04278468470877
-1.12355819707895
-0.903163793466103
-1.04593752600814
-0.940608493489337
-0.919299170031954
-1.10800199186325
-0.960273971678271
-1.0187928851504
-1.04230831877009
-0.950604656347737
-1.12774218434047
-0.93190494233458
-0.943964162071663
-1.02492635287179
-0.936394047319956
-1.05828367305326
