--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED_rscpu_vector_way2.txt
@@ -1,198 +1,0 @@
-"Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED"
-1.5637323788338
-0.978530062323946
-0.57522040242742
-1.6393529400644
-0.982462331507937
-0.840194848979438
-0.394033537718906
-1.31519280092257
-0.577035295896954
-0.390000441119941
-0.231903054440502
-0.605166144674737
-1.63592480795528
-1.32174658289589
-0.613131510457693
-2.33637285978057
-8.51312363661952
-3.499542607839
-1.10140032369291
-0.870874674547885
-0.613327703891352
-2.36014355077053
-3.54964464147491
-1.73583843501513
-0.57814369150658
-0.45908099359651
-0.344803321370769
-0.958693969460277
-1.06649778340722
-0.58883963127155
-0.255295193863908
-0.168601787347829
-0.133417774963057
-0.354654844838505
-0.877630004925762
-0.477095207937513
-0.187741890085145
-0.186897473787911
-0.119625642108226
-0.338329463091971
-0.602350292027303
-0.343958905073535
-0.153120821898529
-0.147491379916966
-0.103018788262613
-0.189149250580536
-2.30272324255858
-0.982337625782844
-0.335233270002111
-0.355780733234818
-0.217577932587432
-0.928013510660756
-0.947151796838907
-0.758886197869272
-0.497487085439095
-1.41238397894472
-0.770981786619255
-0.69213646736011
-0.412593971804959
-1.28190841511389
-0.479231706121992
-0.41225798322857
-0.270134815416276
-0.792261063123854
-1.41809578474332
-1.29881984012544
-0.783189371561367
-3.77247973568899
-0.822209603984122
-0.602978714537252
-1.10384346901991
-0.600822037694961
-0.658442816807493
-1.06413248455423
-1.12029985101218
-1.06024108938227
-1.96702993300757
-0.760071639424336
-0.294981700799218
-0.58168097803388
-0.657879274827804
-1.4355351365892
-1.34646722779496
-0.299059201631
-0.184889178341108
-0.265164975966814
-0.320720924799841
-0.566390349914698
-0.532750968052498
-0.575564726786207
-0.267968257788664
-0.533260655656471
-0.506629478348895
-1.06945201503578
-1.08283131464007
-0.673934434352945
-0.317025689671039
-0.543581829636918
-0.727324210869088
-1.30441800046721
-1.21649688878192
-1.42139130557896
-0.547022220963734
-1.04536927574807
-1.31639565916057
-2.96179466668554
-2.69561031551078
-1.35207379143866
-0.556706285439216
-1.07684248529339
-1.25497830288186
-2.6524142910741
-3.05532234201454
-0.557708458308338
-0.740731853629274
-0.79877024595081
-0.579034193161368
-0.408428314337133
-0.914307138572286
-0.739652069586083
-1.06790641871626
-1.2254199160168
-0.869496100779844
-0.540566886622675
-1.37672465506899
-0.764892021595681
-1.20260947810438
-1.79149670065987
-1.14173665266947
-0.761787642471506
-1.99517096580684
-0.566076784643071
-0.841151769646071
-1.15428914217157
-0.972075584883023
-0.557843431313737
-1.5494901019796
-0.39128674265147
-0.54002699460108
-0.770965806838632
-0.534763047390522
-0.392096580683863
-0.883533293341332
-0.952369526094781
-1.40938812237552
-1.93470305938812
-1.5497600479904
-0.864772045590882
-2.65896820635873
-2.8215023683123
-0.94503608099946
-0.753918046857615
-1.98594907333821
-0.937533776266014
-0.455351750007947
-0.350192326032362
-0.789776520329338
-0.775153383984487
-0.33518771656547
-0.274660647868519
-0.602727532822583
-1.9476746034269
-0.805671233747655
-0.624217185364148
-1.59544775407699
-1.63146245427646
-0.972577526152699
-0.722816336991216
-1.6021993895762
-0.958411966170406
-0.781901633233149
-0.440437082081033
-1.21301926795741
-0.73166981198015
-0.440903054448871
-0.342396495887794
-0.705575359381189
-1.61366230982503
-1.1882295379884
-0.713310500687309
-1.94142727336269
-0.872338847266523
-0.642187770442351
-0.712607867863413
-1.27527631481344
-0.653660707662636
-0.721509284672255
-0.599164255866281
-1.26499023316767
-0.710135252083179
-0.59313107336251
-0.674232870954182
-1.10872091585688
-1.27982592784907
-1.27260588977079
-1.11406176594219
-2.50555102242663
