--- results_examples/Saccharomyces_cerevisiae.CDS.fa.codon_counts.merged.FIXED_sdevs.txt
@@ -1,50 +1,0 @@
-           GCA        GCC        GCG        GCT
-GCA   7.686571 -11.465556   1.719912 -14.603882
-GCC -11.077783   3.093353  -3.690202  -3.624453
-GCG   1.965255  -4.319967   3.746540 -11.151449
-GCT -14.872322  -3.038529 -10.179594   5.385028
-           AGA         AGG         CGA         CGC           CGG         CGT
-AGA   3.878029 -8.49960371 -9.77743712 -12.7756411 -10.476738300  -6.5984385
-AGG  -6.872431  1.93499011 -0.02477794  -2.7921298  -0.006933527 -10.2764822
-CGA -11.647065  0.82027347  7.77909246   0.1101982   2.730793204  -3.3081073
-CGC -12.384235 -1.25031472  2.90136461   6.9924686   3.690161736   0.7908914
-CGG -11.232676 -0.09305286  6.15921691   8.9480190   7.394145827  -5.3165035
-CGT  -8.830041 -8.93506408 -5.17407125   2.6079724  -1.786893586  11.1013820
-           GGA         GGC         GGG       GGT
-GGA   8.617695   0.5642447   4.2008900 -23.36939
-GGC   1.912802   4.1597661   0.7486052 -17.55219
-GGG   1.640410   0.6982309   3.2835952 -13.84253
-GGT -22.933626 -16.1809331 -14.7587433  20.53521
-          ATA        ATC         ATT
-ATA  13.83906 -14.040293 -15.0177226
-ATC -14.43778   7.384145  -5.8833605
-ATT -12.66849  -6.460875   0.2450823
-           CTA       CTC        CTG       CTT        TTA        TTG
-CTA  2.7662595 -3.165142 -0.1884080 -3.661634 -0.5175876  -6.573966
-CTC -2.5186507 11.991263  2.6202843  3.862703 -6.0021608  -9.417727
-CTG -0.9036737  3.118285  9.0969876 -5.239133 -6.5089101  -4.804768
-CTT -1.9350964  3.259586 -0.7915904  8.266788 -4.3404164 -10.530606
-TTA -1.5824174 -8.201548 -8.5335910 -3.413011  6.8902539  -6.760610
-TTG -6.1506204 -6.689166 -5.3097663 -7.542215 -9.1347585  14.003290
-           AGC        AGT        TCA       TCC        TCG        TCT
-AGC  13.064987   8.266873 -6.8840767 -6.939433  1.1188480 -12.889368
-AGT   8.151682   9.147157 -3.4262448 -5.963585 -4.7472848 -12.369482
-TCA -10.044358  -5.200411  8.2773930 -7.041808 -0.4285444  -2.881516
-TCC  -8.342787  -8.519270 -6.0508777  5.934674 -0.7843306   3.207953
-TCG  -1.226401  -4.808682  0.4838044 -3.449150  5.9363337  -4.773979
-TCT  -9.669102 -10.062292 -6.6320316  3.227303 -6.4648365   9.218045
-           CCA          CCC       CCG       CCT
-CCA   3.040494 -12.40629152 -9.819201 -8.857431
-CCC -13.067613   4.09063443  2.419478 -4.184182
-CCG  -7.724107   0.09463201  2.036009 -5.709773
-CCT -11.359563  -2.59674544 -3.303064 -2.165700
-           ACA        ACC       ACG         ACT
-ACA   4.735030 -11.618715  1.476702 -11.9828681
-ACC -13.077523   4.276175 -7.365768   1.0743771
-ACG   2.589292  -7.295692  4.023313  -9.8082084
-ACT -11.030914  -1.361016 -8.882220   0.5706565
-             GTA        GTC         GTG        GTT
-GTA   7.33153326 -11.751400   0.3874841 -12.096846
-GTC -10.37854982   5.421614  -6.4373031  -2.575402
-GTG   0.08553048  -7.205691   6.7197048 -12.239188
-GTT -11.69362489  -1.872959 -11.7368539   2.803035
