--- results_examples/Saccharomyces_cerevisiae.CDS.fa.paired_codons
@@ -1,197 +1,0 @@
-Ala	GCA	GCA	15509
-Ala	GCA	GCC	9744
-Ala	GCA	GCG	5723
-Ala	GCA	GCT	16225
-Ala	GCC	GCA	9705
-Ala	GCC	GCC	8333
-Ala	GCC	GCG	3868
-Ala	GCC	GCT	13109
-Ala	GCG	GCA	5705
-Ala	GCG	GCC	3908
-Ala	GCG	GCG	2300
-Ala	GCG	GCT	6081
-Ala	GCT	GCA	16259
-Ala	GCT	GCC	13044
-Ala	GCT	GCG	6002
-Ala	GCT	GCT	23172
-Arg	AGA	AGA	30245
-Arg	AGA	AGG	12611
-Arg	AGA	CGA	3789
-Arg	AGA	CGC	3118
-Arg	AGA	CGG	2140
-Arg	AGA	CGT	8181
-Arg	AGG	AGA	12433
-Arg	AGG	AGG	6167
-Arg	AGG	CGA	2092
-Arg	AGG	CGC	1695
-Arg	AGG	CGG	1222
-Arg	AGG	CGT	3490
-Arg	CGA	AGA	3913
-Arg	CGA	AGG	2054
-Arg	CGA	CGA	907
-Arg	CGA	CGC	667
-Arg	CGA	CGG	544
-Arg	CGA	CGT	1191
-Arg	CGC	AGA	3094
-Arg	CGC	AGG	1631
-Arg	CGC	CGA	599
-Arg	CGC	CGC	664
-Arg	CGC	CGG	524
-Arg	CGC	CGT	1264
-Arg	CGG	AGA	2179
-Arg	CGG	AGG	1225
-Arg	CGG	CGA	474
-Arg	CGG	CGC	425
-Arg	CGG	CGG	366
-Arg	CGG	CGT	773
-Arg	CGT	AGA	8385
-Arg	CGT	AGG	3406
-Arg	CGT	CGA	1260
-Arg	CGT	CGC	1202
-Arg	CGT	CGG	672
-Arg	CGT	CGT	3297
-Gly	GGA	GGA	8457
-Gly	GGA	GGC	6884
-Gly	GGA	GGG	4279
-Gly	GGA	GGT	12662
-Gly	GGC	GGA	6776
-Gly	GGC	GGC	6180
-Gly	GGC	GGG	3681
-Gly	GGC	GGT	11597
-Gly	GGG	GGA	4442
-Gly	GGG	GGC	3684
-Gly	GGG	GGG	2412
-Gly	GGG	GGT	6993
-Gly	GGT	GGA	12611
-Gly	GGT	GGC	11446
-Gly	GGT	GGG	7074
-Gly	GGT	GGT	33684
-Ile	ATA	ATA	17537
-Ile	ATA	ATC	12815
-Ile	ATA	ATT	23895
-Ile	ATC	ATA	12861
-Ile	ATC	ATC	14044
-Ile	ATC	ATT	22614
-Ile	ATT	ATA	23544
-Ile	ATT	ATC	22697
-Ile	ATT	ATT	41955
-Leu	CTA	CTA	5965
-Leu	CTA	CTC	2347
-Leu	CTA	CTG	4517
-Leu	CTA	CTT	5289
-Leu	CTA	TTA	11155
-Leu	CTA	TTG	10611
-Leu	CTC	CTA	2315
-Leu	CTC	CTC	1451
-Leu	CTC	CTG	2103
-Leu	CTC	CTT	2488
-Leu	CTC	TTA	4293
-Leu	CTC	TTG	4369
-Leu	CTG	CTA	4565
-Leu	CTG	CTC	2081
-Leu	CTG	CTG	4185
-Leu	CTG	CTT	4266
-Leu	CTG	TTA	8204
-Leu	CTG	TTG	8451
-Leu	CTT	CTA	5163
-Leu	CTT	CTC	2517
-Leu	CTT	CTG	3976
-Leu	CTT	CTT	5708
-Leu	CTT	TTA	10331
-Leu	CTT	TTG	9849
-Leu	TTA	CTA	11266
-Leu	TTA	CTC	4445
-Leu	TTA	CTG	8393
-Leu	TTA	CTT	10237
-Leu	TTA	TTA	23244
-Leu	TTA	TTG	20816
-Leu	TTG	CTA	10567
-Leu	TTG	CTC	4181
-Leu	TTG	CTG	8498
-Leu	TTG	CTT	9547
-Leu	TTG	TTA	21155
-Leu	TTG	TTG	23978
-Pro	CCA	CCA	22189
-Pro	CCA	CCC	7373
-Pro	CCA	CCG	6096
-Pro	CCA	CCT	15317
-Pro	CCC	CCA	7432
-Pro	CCC	CCC	3581
-Pro	CCC	CCG	2636
-Pro	CCC	CCT	6336
-Pro	CCG	CCA	5929
-Pro	CCG	CCC	2754
-Pro	CCG	CCG	2160
-Pro	CCG	CCT	4909
-Pro	CCT	CCA	15618
-Pro	CCT	CCC	6211
-Pro	CCT	CCG	4740
-Pro	CCT	CCT	12547
-Ser	AGC	AGC	4132
-Ser	AGC	AGT	5480
-Ser	AGC	TCA	5667
-Ser	AGC	TCC	4194
-Ser	AGC	TCG	2899
-Ser	AGC	TCT	7056
-Ser	AGT	AGC	5488
-Ser	AGT	AGT	7912
-Ser	AGT	TCA	8910
-Ser	AGT	TCC	6232
-Ser	AGT	TCG	4001
-Ser	AGT	TCT	10442
-Ser	TCA	AGC	5918
-Ser	TCA	AGT	9079
-Ser	TCA	TCA	13273
-Ser	TCA	TCC	8552
-Ser	TCA	TCG	5712
-Ser	TCA	TCT	14334
-Ser	TCC	AGC	4290
-Ser	TCC	AGT	6442
-Ser	TCC	TCA	8459
-Ser	TCC	TCC	7202
-Ser	TCC	TCG	3962
-Ser	TCC	TCT	11482
-Ser	TCG	AGC	3026
-Ser	TCG	AGT	4005
-Ser	TCG	TCA	5644
-Ser	TCG	TCC	4133
-Ser	TCG	TCG	2905
-Ser	TCG	TCT	6407
-Ser	TCT	AGC	6774
-Ser	TCT	AGT	10200
-Ser	TCT	TCA	14782
-Ser	TCT	TCC	11480
-Ser	TCT	TCG	6546
-Ser	TCT	TCT	19700
-Thr	ACA	ACA	17506
-Thr	ACA	ACC	10284
-Thr	ACA	ACG	7851
-Thr	ACA	ACT	17315
-Thr	ACC	ACA	10436
-Thr	ACC	ACC	8390
-Thr	ACC	ACG	4731
-Thr	ACC	ACT	12750
-Thr	ACG	ACA	7756
-Thr	ACG	ACC	4726
-Thr	ACG	ACG	3674
-Thr	ACG	ACT	7654
-Thr	ACT	ACA	17192
-Thr	ACT	ACC	13016
-Thr	ACT	ACG	7571
-Thr	ACT	ACT	20832
-Val	GTA	GTA	8820
-Val	GTA	GTC	6609
-Val	GTA	GTG	7180
-Val	GTA	GTT	12940
-Val	GTC	GTA	6493
-Val	GTC	GTC	7295
-Val	GTC	GTG	5997
-Val	GTC	GTT	12867
-Val	GTG	GTA	7205
-Val	GTG	GTC	6058
-Val	GTG	GTG	6817
-Val	GTG	GTT	11264
-Val	GTT	GTA	12894
-Val	GTT	GTC	12790
-Val	GTT	GTG	11210
-Val	GTT	GTT	25333
