--- results_examples/Saccharomyces_cerevisiae.CDS.fa.total_codons
@@ -1,41 +1,0 @@
-Ala	GCA	49169
-Ala	GCC	36509
-Ala	GCG	18758
-Ala	GCT	60888
-Arg	CGA	9715
-Arg	CGC	8258
-Arg	CGG	5792
-Arg	CGT	19143
-Arg	AGA	63169
-Arg	AGG	28460
-Gly	GGT	67526
-Gly	GGG	18369
-Gly	GGC	29617
-Gly	GGA	33972
-Ile	ATT	91241
-Ile	ATC	51253
-Ile	ATA	56141
-Leu	CTT	38489
-Leu	TTG	79741
-Leu	CTG	32446
-Leu	CTC	17519
-Leu	TTA	80235
-Leu	CTA	40807
-Pro	CCT	41144
-Pro	CCG	16552
-Pro	CCC	21058
-Pro	CCA	53715
-Ser	AGT	44212
-Ser	TCT	71104
-Ser	TCG	26681
-Ser	AGC	30389
-Ser	TCC	42870
-Ser	TCA	58166
-Thr	ACT	60844
-Thr	ACG	24772
-Thr	ACC	37813
-Thr	ACA	54938
-Val	GTT	64798
-Val	GTG	32562
-Val	GTC	33983
-Val	GTA	37107
